removed and moved some controllers
