+++ Python/example_controllers/objects_and_scenes/room.png
